feat: add settings page route
